+++ switchbot_client/__init__.py
@@ -1,7 +1,9 @@
-from switchbot_client.client import SwitchBotAPIClient, SwitchBotAPIResponse
+from switchbot_client.api import SwitchBotAPIClient, SwitchBotAPIResponse
+from switchbot_client.client import SwitchBotClient
 from switchbot_client.enums import ControlCommand, DeviceType, RemoteType
 
 __all__ = [
+    "SwitchBotClient",
     "SwitchBotAPIClient",
     "SwitchBotAPIResponse",
     "DeviceType",
