+++ switchbot_client/devices/__init__.py
@@ -1,3 +1,4 @@
 from .base import *  # noqa
+from .factory import *  # noqa
 from .physical import *  # noqa
 from .remote import *  # noqa
